#2227 moved link handling to core reader activity
