+++ core/src/main/java/org/kiwix/kiwixmobile/core/main/CoreMainActivity.kt
@@ -21,14 +21,15 @@ import android.content.Intent
 import android.net.Uri
 import android.view.ActionMode
 import android.view.MenuItem
+import android.widget.Toast
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.widget.Toolbar
+import androidx.core.content.ContextCompat
 import com.google.android.material.navigation.NavigationView
 import org.kiwix.kiwixmobile.core.R
 import org.kiwix.kiwixmobile.core.base.BaseActivity
 import org.kiwix.kiwixmobile.core.base.BaseFragmentActivityExtensions
 import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.intent
-import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.openExternalUrl
 import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.start
 import org.kiwix.kiwixmobile.core.extensions.browserIntent
 import org.kiwix.kiwixmobile.core.help.HelpActivity
@@ -37,7 +38,9 @@ import org.kiwix.kiwixmobile.core.page.history.HistoryActivity
 import org.kiwix.kiwixmobile.core.utils.AlertDialogShower
 import org.kiwix.kiwixmobile.core.utils.BOOKMARK_CHOSEN_REQUEST
 import org.kiwix.kiwixmobile.core.utils.EXTRA_EXTERNAL_LINK
+import org.kiwix.kiwixmobile.core.utils.KiwixDialog
 import org.kiwix.kiwixmobile.core.utils.REQUEST_HISTORY_ITEM_CHOSEN
+import org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil
 import javax.inject.Inject
 
 abstract class CoreMainActivity : BaseActivity(), WebViewProvider,
@@ -116,6 +119,45 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider,
 
   abstract fun navigationDrawerIsOpen(): Boolean
   abstract fun closeNavigationDrawer()
+
+  private fun urlIntentIsValid(intent: Intent) =
+    (intent.hasExtra(EXTRA_EXTERNAL_LINK) && intent.getBooleanExtra(EXTRA_EXTERNAL_LINK, false))
+
+  private fun requestOpenLink(
+    alertDialogShower: AlertDialogShower,
+    intent: Intent,
+    sharedPreferenceUtil: SharedPreferenceUtil
+  ) {
+    alertDialogShower.show(
+      KiwixDialog.ExternalLinkPopup, { ContextCompat.startActivity(this, intent, null) },
+      {}, {
+        sharedPreferenceUtil.putPrefExternalLinkPopup(false)
+        ContextCompat.startActivity(this, intent, null)
+      })
+  }
+
+  private fun openExternalUrl(
+    sharedPreferenceUtil: SharedPreferenceUtil,
+    alertDialogShower: AlertDialogShower,
+    intent: Intent
+  ) {
+    if (intent.resolveActivity(packageManager) != null) {
+      // Show popup with warning that this url is external and could lead to additional costs
+      // or may event not work when the user is offline.
+      if (urlIntentIsValid(intent) && sharedPreferenceUtil.prefExternalLinkPopup) {
+        requestOpenLink(alertDialogShower, intent, sharedPreferenceUtil)
+      } else {
+        openLink(intent)
+      }
+    } else {
+      val error = getString(R.string.no_reader_application_installed)
+      Toast.makeText(this, error, Toast.LENGTH_LONG).show()
+    }
+  }
+
+  private fun openLink(intent: Intent) {
+    startActivity(intent)
+  }
 
   private fun openSupportKiwixExternalLink() {
     openExternalUrl(
